Division del codigo en modulos
Se dividio el codigo en modulos y se hizo pequeños cambios en cuanto a funcionalidad e interfaz
