+++ clase-12/tarea/tarea-exchanges/index.html
@@ -6,6 +6,7 @@
         <title>API exchanges</title>
         <link rel="stylesheet" type="text/css" href="node_modules\bootstrap\dist\css\bootstrap.css">
         <link rel="stylesheet" type="text/css" href="src/style.css">
+        <link rel="shortcut icon" href="">
     </head>
 
 
@@ -19,7 +20,7 @@
                         <label class="input-group-text">Base </label>
                     </div>
                     <select class="custom-select" id="listado-bases">
-                        <option selected>Elige una moneda</option>
+                        <option selected id="selected-base">Cargando bases...</option>
                     </select>
                 </div>
                 
@@ -44,6 +45,6 @@
 
         </div>
         <script type="text/javascript" src="node_modules/jquery/dist/jquery.min.js"></script>
-        <script src="src/index.js"></script>
+        <script src="./src/index(clase-13).js" type="module"></script>
     </body>
 </html>
