+++ clase-12/tarea/tarea-exchanges/src/index(clase-13).js
@@ -1,1 +1,34 @@
 /// <reference types="jquery"/>
+
+import { obtenerNombresBases } from './exchangesAPI.js'
+import {
+  añadirBases,
+  ObtenerURL,
+  VaciarDatosPrevios,
+  AgregarTarjetasCambios,
+  errorFetch,
+  agragarGifCargando
+} from './interfaz.js'
+
+function configurarSelector () {
+  obtenerNombresBases(añadirBases)
+}
+
+configurarSelector()
+
+$('#convertir').click(async () => {
+  VaciarDatosPrevios()
+  agragarGifCargando()
+
+  $.ajax({
+    method: 'GET',
+    url: ObtenerURL(),
+    success: respuesta => {
+      $('#gif').detach()
+      AgregarTarjetasCambios(respuesta)
+    },
+    error: () => {
+      $('form').append(errorFetch())
+    }
+  })
+})
